refactor : 패키지 구조 수정 model, component 로 이원화되어있던것을 > model 로 단일화
